+++ internal/bundler_tests/bundler_tsconfig_test.go
@@ -1199,6 +1199,140 @@ func TestTsconfigJsonNodeModulesImplicitFile(t *testing.T) {
 	})
 }
 
+func TestTsconfigJsonNodeModulesTsconfigPathExact(t *testing.T) {
+	tsconfig_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/app/entry.tsx": `
+				console.log(<div/>)
+			`,
+			"/Users/user/project/src/tsconfig.json": `
+				{
+					"extends": "foo"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/package.json": `
+				{
+					"tsconfig": "over/here.json"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/over/here.json": `
+				{
+					"compilerOptions": {
+						"jsx": "react",
+						"jsxFactory": "worked"
+					}
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/app/entry.tsx"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+	})
+}
+
+func TestTsconfigJsonNodeModulesTsconfigPathImplicitJson(t *testing.T) {
+	tsconfig_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/app/entry.tsx": `
+				console.log(<div/>)
+			`,
+			"/Users/user/project/src/tsconfig.json": `
+				{
+					"extends": "foo"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/package.json": `
+				{
+					"tsconfig": "over/here"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/over/here.json": `
+				{
+					"compilerOptions": {
+						"jsx": "react",
+						"jsxFactory": "worked"
+					}
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/app/entry.tsx"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+	})
+}
+
+func TestTsconfigJsonNodeModulesTsconfigPathDirectory(t *testing.T) {
+	tsconfig_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/app/entry.tsx": `
+				console.log(<div/>)
+			`,
+			"/Users/user/project/src/tsconfig.json": `
+				{
+					"extends": "foo"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/package.json": `
+				{
+					"tsconfig": "over/here"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/over/here/tsconfig.json": `
+				{
+					"compilerOptions": {
+						"jsx": "react",
+						"jsxFactory": "worked"
+					}
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/app/entry.tsx"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+	})
+}
+
+func TestTsconfigJsonNodeModulesTsconfigPathBad(t *testing.T) {
+	tsconfig_suite.expectBundled(t, bundled{
+		files: map[string]string{
+			"/Users/user/project/src/app/entry.tsx": `
+				console.log(<div/>)
+			`,
+			"/Users/user/project/src/tsconfig.json": `
+				{
+					"extends": "foo"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/package.json": `
+				{
+					"tsconfig": "over/here.json"
+				}
+			`,
+			"/Users/user/project/src/node_modules/foo/tsconfig.json": `
+				{
+					"compilerOptions": {
+						"jsx": "react",
+						"jsxFactory": "THIS SHOULD NOT BE LOADED"
+					}
+				}
+			`,
+		},
+		entryPaths: []string{"/Users/user/project/src/app/entry.tsx"},
+		options: config.Options{
+			Mode:          config.ModeBundle,
+			AbsOutputFile: "/Users/user/project/out.js",
+		},
+		expectedScanLog: `Users/user/project/src/tsconfig.json: WARNING: Cannot find base config file "foo"
+`,
+	})
+}
+
 func TestTsconfigJsonInsideNodeModules(t *testing.T) {
 	tsconfig_suite.expectBundled(t, bundled{
 		files: map[string]string{